Scottish Tax option added
Incorporated the Scottish Tax rates,separated out a calculation function for this, and enabled the Scottish Tax checkbox to modify the outputs to use Scottish Tax Rates.

Found a bug where reduced Personal Allowance (on both sets of Tax Rates) was being taxed at Basic Rate not Higher Rate - so have fixed this.