Move misc scripts into scripts.js, edit placeholder inputIncome.
Misc scripts for interactive elements, toggling classes etc. have been moved to scripts.js, and the remaining content of take_home_pay_calculator.js is more relevant.
The default value of inputIncome (25000) has been replaced with a placeholder attribute which takes place of a blank value in the field. The pie chart and outputs reflect the input value of 25000 when the input is blank.